+++ arviz/wrappers/__init__.py
@@ -1,5 +1,5 @@
 """Sampling wrappers."""
 from .base import SamplingWrapper
-from .wrap_pystan import PyStanSamplingWrapper
+from .wrap_stan import PyStan2SamplingWrapper, PyStanSamplingWrapper
 
-__all__ = ["SamplingWrapper", "PyStanSamplingWrapper"]
+__all__ = ["SamplingWrapper", "PyStan2SamplingWrapper", "PyStanSamplingWrapper"]
